+++ macOS/Sources/MPXPrime/MPXPrimeViewModel.swift
@@ -309,6 +309,7 @@ final class MPXPrimeViewModel: ObservableObject {
     var compositeOverBudget: Bool { get { telemetry.compositeOverBudget } set { telemetry.compositeOverBudget = newValue } }
     var compositeClipperGainReductionDBValue: Float { get { telemetry.compositeClipperGainReductionDBValue } set { telemetry.compositeClipperGainReductionDBValue = newValue } }
     var bs412StatusValue: BS412Status { get { telemetry.bs412StatusValue } set { telemetry.bs412StatusValue = newValue } }
+    var nonFiniteInputSamplesValue: UInt64 { get { telemetry.nonFiniteInputSamplesValue } set { telemetry.nonFiniteInputSamplesValue = newValue } }
     var compositeClipperLookaheadGainReductionDBValue: Float { get { telemetry.compositeClipperLookaheadGainReductionDBValue } set { telemetry.compositeClipperLookaheadGainReductionDBValue = newValue } }
     var preEncodeLimiterGainReductionDBValue: Float { get { telemetry.preEncodeLimiterGainReductionDBValue } set { telemetry.preEncodeLimiterGainReductionDBValue = newValue } }
     var safetyLimiterGainReductionDBValue: Float { get { telemetry.safetyLimiterGainReductionDBValue } set { telemetry.safetyLimiterGainReductionDBValue = newValue } }
@@ -1977,6 +1978,7 @@ final class MPXPrimeViewModel: ObservableObject {
         var dacPeak: Float = 0.0
         var compositeClipperGainReductionDB: Float = 0.0
         var bs412 = BS412Status()
+        var nonFiniteInputSamples: UInt64 = 0
         var compositeClipperLookaheadGainReductionDB: Float = 0.0
         var preEncodeAudioLimiterGainReductionDB: Float = 0.0
         var mpxSafetyLimiterGainReductionDB: Float = 0.0
@@ -2120,6 +2122,7 @@ final class MPXPrimeViewModel: ObservableObject {
             dacPeak = meters.dacPeak
             compositeClipperGainReductionDB = meters.compositeClipperGainReductionDB
             bs412 = meters.bs412
+            nonFiniteInputSamples = meters.nonFiniteInputSamples
             compositeClipperLookaheadGainReductionDB = meters.compositeClipperLookaheadGainReductionDB
             preEncodeAudioLimiterGainReductionDB = meters.preEncodeAudioLimiterGainReductionDB
             mpxSafetyLimiterGainReductionDB = meters.mpxSafetyLimiterGainReductionDB
@@ -2200,6 +2203,7 @@ final class MPXPrimeViewModel: ObservableObject {
             compositeOverBudget = false
             compositeClipperGainReductionDBValue = 0.0
             bs412StatusValue = BS412Status()
+            nonFiniteInputSamplesValue = 0
             compositeClipperLookaheadGainReductionDBValue = 0.0
             preEncodeLimiterGainReductionDBValue = 0.0
             safetyLimiterGainReductionDBValue = 0.0
@@ -2320,6 +2324,7 @@ final class MPXPrimeViewModel: ObservableObject {
         self.compositeOverBudget = compositeOverBudget
         compositeClipperGainReductionDBValue = compositeClipperGainReductionDB
         bs412StatusValue = bs412
+        nonFiniteInputSamplesValue = nonFiniteInputSamples
         compositeClipperLookaheadGainReductionDBValue = compositeClipperLookaheadGainReductionDB
         preEncodeLimiterGainReductionDBValue = preEncodeAudioLimiterGainReductionDB
         safetyLimiterGainReductionDBValue = mpxSafetyLimiterGainReductionDB